fix(backend): resolve 3 Docker smoke-test router failures blocking CI
Three feature routers crashed under AUTOBOT_FEATURE_ROUTERS_STRICT=1:

1. api.playwright — AiProposeRegionsResponse missing from schemas_code
   (already fixed in prior commit, listed here for context)

2. api.code_intelligence — AntiPatternDetector.GOD_CLASS_LINE_THRESHOLD
   not present on the code_analysis.src canonical class; add alias to
   GOD_CLASS_LOC_THRESHOLD (same value, 500 lines).

3. api.knowledge_test — .dockerignore pattern **/*_test.py silently
   excluded this real router from the Docker image. Moved implementation
   to api/knowledge_eval.py; left api/knowledge_test.py as a thin
   re-export shim for existing test-suite imports; updated feature_routers
   to register api.knowledge_eval.
